Add files via upload
Main code for a custom water quality monitoring system. This specific system measures pH, temperature, and relative conductivity of the water. In addition, this system has a alarm that is dedicated to alerting the user that certain parameters are out of certain thresholds set, and it also records/displays all data for the user.